Changed 'int nmiss' to 'size_t nmiss'.

+++ src/Cond.cc
@@ -35,7 +35,7 @@ void *Cond(void *argument)
   int nrecs, nrecs2, nvars = 0, nlev;
   int varID, levelID;
   int offset;
-  int nmiss1, nmiss2, nmiss3;
+  size_t nmiss1, nmiss2, nmiss3;
   int i;
   double missval1 = -9.E33;
   double missval2 = -9.E33;
