+++ src/Input.cc
@@ -245,7 +245,7 @@ void *Input(void *argument)
 	  pstreamDefRecord(streamID, varID, levelID);
 
           int offset = gridsize*levelID;
-	  int nmiss = 0;
+	  size_t nmiss = 0;
 	  for ( i = 0; i < gridsize; ++i )
 	    if ( DBL_IS_EQUAL(array[offset+i], missval) ) nmiss++;
